Refactoring for Clean Architecture
