+++ app/src/main/java/com/trainingtimer/database/TrainingDao.kt
@@ -5,7 +5,7 @@ import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.Query
 import androidx.room.Update
-import com.trainingtimer.Training
+import com.trainingtimer.domain.Training
 import java.util.*
 
 @Dao
