+++ app/src/main/java/com/trainingtimer/database/TrainingDatabase.kt
@@ -3,7 +3,7 @@ package com.trainingtimer.database
 import androidx.room.Database
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
-import com.trainingtimer.Training
+import com.trainingtimer.domain.Training
 
 @Database(entities = [ Training::class], version = 1)
 @TypeConverters(TrainingTypeConverters::class)
