+++ app/src/main/res/layout/time_picker_dialog.xml
@@ -6,7 +6,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:gravity="center_horizontal">
 
-    <com.trainingtimer.TimePicker
+    <com.trainingtimer.domain.TimePicker
         android:id="@+id/timePicker"
         android:layout_gravity="center_horizontal"
         android:layout_width="wrap_content"
